requirements.txt
Django==1.11.6
django-storages==1.4.1
raven==6.9.0
enum34==1.1.6
requests-kerberos==0.11.0
mock==2.0.0
django-fine-uploader-s3==0.1.14
boto3==1.4.7
boto==2.41.0
deepdiff==3.3.0
gspread==0.6.2
oauth2client==4.1.2
logentries==0.17
tblib==1.3.2
redis==2.10.6
pynamodb==3.3.0
django-cacheops==4.0.7
schema==0.6.6
redisco==0.1.4 # Dependency from ib_common
google-api-python-client==1.7.6

django-swagger-utils==1.3.23
ib_common==0.1.62
ib_gas_auth==1.0.9

attrs==16.3.0
